requirements.txt
# --- Web & API ---
fastapi
uvicorn
requests
python-dotenv
Pillow

# --- AI Search & Vision ---
duckduckgo-search
transformers
torch --index-url https://download.pytorch.org/whl/cpu
huggingface_hub

# --- Automation & Messaging ---
playwright==1.42.0
patchright==1.42.0
cloudinary
python-telegram-bot
tenacity
